refactor: consolidate integration tests, remove redundancy
Move shared-client tests that only existed in filozzy-mcp into
ghprojects-client (6 tests). Strip filozzy-mcp down to 2 MCP-layer-
specific tests (docstring query validation). Zero overlap between
the two test suites.

Before: 21 + 29 = 50 tests (heavy duplication)
After:  27 +  2 = 29 tests (no duplication)
